Smart Interaction With Data

smart.data is an R6 class that facilitates the manipulation of a data set by providing rules-based operations and taxonomical column reference.

You will need to install related package book.of.utilities

Public fields

data

A data.table =object holding the current state of the data

name

The label to use for the data

id

A unique ID for the object

smart.rules

A list holding user-generated sets of rules

Methods

Public methods

smart.data$print()

smart.data$new()

smart.data$naming.rule()

smart.data$taxonomy.rule()

smart.data$transformation.rule()

smart.data$enforce.rules()

smart.data$set.private()

smart.data$reset()

smart.data$use()

smart.data$cache_mgr()

smart.data$clone()

Method print()

print Prints the contents of $data before invisibly returning the class object

Usage

smart.data$print()

Method new()

Initialize the class object.

Usage

smart.data$new(x, name = "new_data", ...)

Arguments

x

The input data

name

The name for the smart class when used in smart functions

...

Arguments used to initialize the smart cache (see cache_layered). If none are provided, a composite cache is created of types memory and disk, both using defaults (see cache_mem and cache_disk)

Method naming.rule()

$naming.rule provides the prescribed laws for renaming the fields found in self$data. The specified fields found in self$data are renamed when the "law" is enforced.

Usage

smart.data$naming.rule(..., show = FALSE)

Arguments

...

Key-value pairs denoting the naming scheme: new_name = old_name. Items with no key will be used as its own name. dots_list supported.

show

(logical) When TRUE the structure of the rule is printed to console

Returns

Before invisibly returning the class object, a list is saved to self$smart.rules$for_naming containing derived objects used by attribute "law": the expression representing the actions to take when enforced

Method taxonomy.rule()

$taxonomy.rule() sets class object $smart.rules$for_usage which is referenced by method $use(). $taxonomy.rule creates a mapping between the usage and data type of a field, labeled by specific terms, and the names of the fields that participate.

Parameter term.map is a data.table object with the following required fields:

term

The name of the taxonomy term. Defaults are c("identifier", "event.date", "flag", "demographic", "category", "join.key")

desc

A description for each term's interpretation or context for usage

rule

A quoted expression containing the rules to be enforced on fields mapped to the current term. The variable for the field must be expressed as 'x' or '.x' and the field name as 'y' or '.y': the law that is enforced exists in the context of imap

fields

The fields in $data mapped to the term

Usage

smart.data$taxonomy.rule(
  term.map = NULL,
  update = FALSE,
  show = FALSE,
  chatty = FALSE
)

Arguments

term.map

(NULL) A 2D tabular object that maps terms to rules -> X ~ term<character[]> + desc<character[]> + rule<language[]> (see section 'Term Map')

update

(logical|TRUE) Should the existing map be updated via interactive selection?

show

(logical|FALSE) Should the rule be shown?

chatty

(logical|FALSE) Should additional information be printed to the console?

new.terms

(data.table|NULL) A link[data.table]{data.table} object of additional taxonomy rules to add to existing ones.

Method transformation.rule()

$transformation.rule() updates class field $data. Rules should be provided as quoted expressions. If given as key-value pairs, each key will print to console as a message during enforcement (regardless of the value of argument chatty). Any loops or other control-of-flow should be defined here, and the object of transformation must use the symbol this.data and scoped assignment operator <<- (e.g. this.data <<- 'expression')

Usage

smart.data$transformation.rule(
  ...,
  set.active = 0,
  show = FALSE,
  chatty = FALSE,
  update = FALSE
)

Arguments

...

Quoted expression containing the code that execute the desired transformations If names are provided (e.g., does_this = quote(...)), the name will be printed to console if the rule enforcement argument chatty is |codeTRUE

set.active

(integer[]) When provided, the indices of self$smart.rules$for_transformation that should be set to 'active': other indices are set to 'inactive'

show

(logical|FALSE) Should the rule be shown?

chatty

(logical|FALSE) Should additional information be printed to the console?

update

(logical|FALSE) Should provided rules function as updates?

Method enforce.rules()

$enforce.rules operates on rules saved in $smart.rules and evaluates the quoted "law" on $data. Custom rules can be directly added to class member $smart.rules: they must have an attribute "law" containing a quoted expression referencing object .data and rule. The rule should contain any objects referenced by the quoted expression.

Usage

smart.data$enforce.rules(..., chatty = FALSE)

Arguments

...

(string[]) The names of the rules to enforce

chatty

(logical|FALSE) Should additional execution information be printed to console?

Returns

Invisibly, the class object with member $data modified according to the rules enforced

Method set.private()

$set.private operates on the private class list accessed internally as $private using key-value pairs (i.e. key = value)

Usage

smart.data$set.private(...)

Arguments

...

A named list: existing names in private indicate elements to overwrite

Returns

Invisibly, the class object with member $data modified according to the rules enforced

Method reset()

$reset restores the object $data to the original value supplied when the class was initiated. Rules are set to a state of "pending", and transformation rules are set as inactive (active = FALSE)

Usage

smart.data$reset(chatty = TRUE)

Arguments

chatty

(logical | TRUE) When TRUE, a confirmation dialog is invoked.

Returns

Invisibly, the class object with member $data modified to the original state

Method use()

$use takes as input taxonomy terms and returns the fields in $data mapped to each term supplied. The idea here is to provide a semantic method for retrieving data.

When no arguments are supplied, the default is to return the fields mapped to existing terms. Using retain = '' returns all values making it equivalent to smart_obj$data

Usage

smart.data$use(
  ...,
  subset = TRUE,
  retain = NULL,
  omit = NULL,
  show = FALSE,
  chatty = FALSE
)

Arguments

...

Taxonomy terms as defined by self$transformation.rule: can be object names or strings

subset

A list of quoted expressions that will row-wise filter the data using data.table syntax

retain

A vector of strings or symbols denoting the fields to retain in the output. Pattern-matching is supported

omit

A vector of strings or symbols denoting the fields to omit from the output

show

(logical|FALSE) Should the rule be shown?

chatty

(logical | TRUE) When TRUE, a confirmation dialog is invoked.

Returns

self$data with columns selected based on the terms supplied

Method cache_mgr()

$cache_mgr adds the current object to the shared "smart" cache space (see cache_layered). The shared cache is layered as 'memory' followed by 'disk.'

Usage

smart.data$cache_mgr(action, chatty = FALSE, ...)

Arguments

action

One of register or unregister

chatty

(logical | TRUE) When TRUE, a confirmation dialog is invoked.

...

Additional arguments to use when initializing a cache object

Method clone()

The objects of this class are cloneable with this method.

Usage

smart.data$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.